changes to the icons.
swapped between the default token 1 and token 2 when opening the dex.

updated the swapCompleted Modal to accommodate both when the message is and is not added.

changed the wording in the messages swap.

changed the settings in adding the polygon swap chain modal - the button just works much better. Also changed the wording of the modal and removed the ability to close it.

App js - added a bit of a hack that checks whether the swapCompletedModal is raised during a transaction, and if it is it updates the allowance to reflect the new purchase.